pyproject.toml
[project]
name = "slowbuild"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=0.3.41",
    "langchain-openai>=0.3.7",
    "langgraph>=0.3.5",
    "chainlit",
    "langchain_community",
    "langchain_qdrant",
    "qdrant_client",    
    "unstructured",
    "python-dotenv",
    "pydantic>=2.0.0"
]